+++ cli/running/process_test.go
@@ -31,6 +31,8 @@ func TestNewInstanceProcess(t *testing.T) {
 	assert.Equal("tmp/run/myapp.instance-1.pid", process.pidFile)
 	assert.Equal("tmp/log", process.logDir)
 	assert.Equal("tmp/log/myapp.instance-1.log", process.logFile)
+	assert.Equal("tmp/run/myapp.instance-1.control", process.consoleSock)
+
 	assert.Equal("tmp/run/myapp.instance-1.notify", process.notifySockPath)
 
 	expEnv := []string{
@@ -69,6 +71,7 @@ func TestNewStateboardProcess(t *testing.T) {
 	assert.Equal("tmp/run/myapp-stateboard.pid", process.pidFile)
 	assert.Equal("tmp/log", process.logDir)
 	assert.Equal("tmp/log/myapp-stateboard.log", process.logFile)
+	assert.Equal("tmp/run/myapp-stateboard.control", process.consoleSock)
 
 	assert.Equal("tmp/run/myapp-stateboard.notify", process.notifySockPath)