+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `--force` option for `start` command to send SIGKILL to instances
+- `cartridge clean` command to remove instance(s) files
 
 ## [2.1.0] - 2020-07-17